CorrigiendoErorres

+++ clinicaOdontologica/src/main/java/com/digitalhouse/backend/integrador/clinicaOdontologica/dao/IDao.java
@@ -6,9 +6,9 @@ public interface IDao<T> {
     //alta, buscarlos, eliminarlos y listarlos
     T registrar(T t);
 
-    T buscarPorId(int id);
+    T buscarPorId(Long id);
 
-    void eliminar(int id);
+    void eliminar(Long id);
 
     List<T> listarTodos();
